requirements.txt
appdirs==1.4.4
asgiref==3.3.1
attrs==20.3.0
audioread==2.1.9
Automat==20.2.0
beautifulsoup4==4.9.3
blinker==1.4
Brotli==1.0.9
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
constantly==15.1.0
crcmod==1.7
cryptography==3.3.2
cssselect==1.1.0
decorator==4.4.2
Flask==1.1.2
h11==0.12.0
h2==4.0.0
hpack==4.0.0
hyperframe==6.0.0
hyperlink==21.0.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.4
incremental==21.3.0
itemadapter==0.2.0
itemloaders==1.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
kaitaistruct==0.9
ldap3==2.8.1
librosa==0.6.3
llvmlite==0.31.0
loguru==0.5.3
lxml==4.6.3
MarkupSafe==1.1.1
mitmproxy==6.0.2
moviepy==1.0.3
msgpack==1.0.2
mysqlclient==2.0.3
natsort==7.1.1
numba==0.48.0
numpy==1.21.0
opencv-python==4.5.2.54
packaging==20.9
pandas==1.3.1
parsel==1.6.0
passlib==1.7.4
Pillow==8.2.0
pooch==1.4.0
proglog==0.1.9
Protego==0.1.16
protobuf==3.14.0
publicsuffix2==2.20191221
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
pydivert==2.1.0
PyMySQL==1.0.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyperclip==1.8.2
python-dateutil==2.8.2
pytz==2021.1
queuelib==1.5.0
requests==2.25.1
resampy==0.2.2
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
scikit-learn==0.24.2
scipy==1.7.0
Scrapy==2.4.1
service-identity==18.1.0
six==1.15.0
sortedcontainers==2.3.0
SoundFile==0.10.3.post1
soupsieve==2.2.1
threadpoolctl==2.1.0
tornado==6.1
tqdm==4.61.1
Twisted==21.2.0
twisted-iocpsupport==1.0.1
urllib3==1.26.4
urwid==2.1.2
w3lib==1.22.0
Werkzeug==1.0.1
win32-setctime==1.0.3
wsproto==1.0.0
zope.interface==5.3.0
zstandard==0.14.1
